refactor: 어노테이션(AuthUser, ExtractDeviceId) 공통 모듈로 이동
