+++ nmnb-application/src/main/kotlin/nmnb/application/global/auth/generator/AuthUserArgumentResolver.kt
@@ -1,6 +1,6 @@
 package nmnb.application.global.auth.generator
 
-import nmnb.application.global.auth.generator.annotation.AuthUser
+import nmnb.common.handler.annotation.AuthUser
 import nmnb.common.response.exception.AuthException
 import nmnb.common.response.status.ErrorStatus
 import nmnb.domain.user.User
