--- nmnb-application/src/main/kotlin/nmnb/application/global/auth/generator/annotation/AuthUser.kt
@@ -1,5 +1,0 @@
-package nmnb.application.global.auth.generator.annotation
-
-@Retention(AnnotationRetention.RUNTIME)
-@Target(AnnotationTarget.VALUE_PARAMETER)
-annotation class AuthUser
